+++ lib/pushpanjali/ticket_page.dart
@@ -9,7 +9,6 @@ import 'package:garuda/pushpanjali/record.dart';
 import 'package:garuda/pushpanjali/tally_cash.dart';
 import 'package:garuda/pushpanjali/tally_upi_card.dart';
 import 'package:garuda/pushpanjali/sevaslot.dart';
-import 'package:flip_card/flip_card.dart';
 
 class TicketTable extends StatefulWidget {
   const TicketTable({super.key});