+++ changelog.md
@@ -1,3 +1,4 @@
 # 1.2.0
 - rename Pushpanjali slot
 - delete Pushpanjali slot by admin
+- user approval auto refresh